Remove BuildCSV job (deprecated)
