+++ deepr/examples/movielens/jobs/__init__.py
@@ -1,6 +1,5 @@
 # pylint: disable=unused-import,missing-docstring
 
 from deepr.examples.movielens.jobs.build_records import BuildRecords
-from deepr.examples.movielens.jobs.build_csv import BuildCSV
 from deepr.examples.movielens.jobs.predict import Predict
 from deepr.examples.movielens.jobs.evaluate import Evaluate
